Tidied away the CtU2015 hack week code, and added read and show methods to the beatbox.Universe class
