+++ beatbox/__init__.py
@@ -1,1 +1,2 @@
 from universe import *
+from ctu2015 import *
